+++ src/components/manifold-data-deprovision-button/manifold-data-deprovision-button.tsx
@@ -18,12 +18,14 @@ interface SuccessMessage {
   message: string;
   resourceLabel: string;
   resourceId: string;
+  ownerId?: string;
 }
 
 interface ErrorMessage {
   message: string;
-  resourceLabel: string;
+  ownerId?: string;
   resourceId?: string;
+  resourceLabel: string;
 }
 
 @Component({ tag: 'manifold-data-deprovision-button' })
@@ -68,12 +70,11 @@ export class ManifoldDataDeprovisionButton {
     this.click.emit({
       resourceId: this.resourceId,
       resourceLabel: this.resourceLabel || '',
+      ownerId: this.ownerId,
     });
 
-    const variables: DeleteResourceMutationVariables = {
-      resourceId: this.resourceId,
-      owner: this.ownerId,
-    };
+    // Note(drew): because we send resourceId here, we DO NOT need owner
+    const variables: DeleteResourceMutationVariables = { resourceId: this.resourceId };
     const { data, errors } = await this.graphqlFetch<DeleteResourceMutation>({
       query: deleteMutation,
       variables,
@@ -84,8 +85,9 @@ export class ManifoldDataDeprovisionButton {
       // success
       const success: SuccessMessage = {
         message: `${data.deleteResource.data.label} successfully deleted`,
-        resourceLabel: data.deleteResource.data.label,
+        ownerId: this.ownerId,
         resourceId: data.deleteResource.data.id,
+        resourceLabel: data.deleteResource.data.label,
       };
       this.success.emit(success);
     }
@@ -94,6 +96,7 @@ export class ManifoldDataDeprovisionButton {
       errors.forEach(({ message }) => {
         const error: ErrorMessage = {
           message,
+          ownerId: this.ownerId,
           resourceLabel: this.resourceLabel || '',
           resourceId: this.resourceId,
         };